+++ web/src/beta/components/fields/CameraField/types.ts
@@ -1,0 +1,5 @@
+import type { Camera } from "@reearth/beta/utils/value";
+
+export type { Camera } from "@reearth/beta/utils/value";
+
+export type RowType = { id: keyof Camera; description?: string; suffix?: string }[];
